Move ShaderSource to a separate pair of files.
