+++ src/apps/graphics_tests/graphics_tests/ShaderSource_tests.cpp
@@ -1,5 +1,6 @@
 #include "catch.hpp"
 
+#include <renderer/ShaderSource.h>
 #include <renderer/Shading.h>
 #include <graphics/ApplicationGlfw.h>
