+++ src/libs/renderer/renderer/Shading.h
@@ -1,108 +1,23 @@
 #pragma once
 
 #include <handy/Guard.h>
-#include <handy/tuple_utils.h>
 
 #include "GL_Loader.h"
 
-// TODO Simplify the inclusions, while moving out ShaderSource.
-#include <filesystem>
-#include <functional>
-#include <map>
-#include <stack>
-
-#include <vector>
 
 namespace ad {
 namespace graphics {
 
-struct SourceMap
-{
-    struct Mapping
-    {
-        std::string mIdentifier;
-        std::size_t mLine;
-    };
 
-    virtual Mapping getLine(std::size_t aCompiledSourceLine) const = 0;
-};
+// Defined in ShaderSource.h
+class ShaderSource;
+struct SourceMap;
 
 
-/// \brief Host the shader code string, and offer preprocessing for #include
-class ShaderSource
-{
-    friend struct ShaderSourceView;
-
-    class InclusionSourceMap : public SourceMap
-    {
-    public:
-        using IdentifierId = std::size_t;
-
-        SourceMap::Mapping getLine(std::size_t aCompiledSourceLine) const override;
-
-        IdentifierId registerSource(std::string_view aIdentifier);
-        void associateLines(IdentifierId aIdentifier,
-                            std::size_t aAssembledLine,
-                            std::size_t aSourceLine);
-        
-    private:
-        struct OriginalLine
-        {
-            std::size_t mIdentifierIndex;
-            std::size_t mLineNumber;
-        };
-
-        std::vector<std::string> mIdentifiers;
-        std::map<std::size_t/*line in assembled output*/, OriginalLine> mMap;
-    };
-
-public:
-    /// \brief The lookup function takes the string as found inside the include directive quotes,
-    /// and returns (unique pointer to) the opened stream and the identifier for the included content.
-    /// \note In the case of including files in nested subfolders, the returned identifier could be the full path
-    /// whereas the input (from the include directive) might be a relative path from the current folder.
-    using Lookup = std::function<
-        std::pair<std::unique_ptr<std::istream>, std::string/*identifier*/>(const std::string &)>;
-
-    static ShaderSource Preprocess(std::istream & aIn,
-                                   const std::string & aIdentifier,
-                                   const Lookup & aLookup);
-
-    static ShaderSource Preprocess(std::istream && aIn,
-                                   const std::string & aIdentifier,
-                                   const Lookup & aLookup)
-    { return Preprocess(aIn, aIdentifier, aLookup); }
-
-    static ShaderSource Preprocess(std::filesystem::path aFile);
-
-    const InclusionSourceMap & getSourceMap() const
-    { return mMap; }
-
-private:
-    struct Input
-    {
-        std::istream & mStream;
-        std::string_view mIdentifier;
-    };
-
-    struct Assembled
-    {
-        std::ostringstream mStream;
-        std::size_t mOutputLine{0};
-        InclusionSourceMap mMapping;
-    };
-
-    static void Preprocess_impl(Input aIn, Assembled & aOut, const Lookup & aLookup);
-
-    ShaderSource(std::string aSource, InclusionSourceMap aMap);
-
-    std::string mSource;
-    InclusionSourceMap mMap;
-};
-
-
-/// \brief Augments the (char *) shader source code (as a string_view) with an
-/// optional name helping to identify the shader associated to the source.
+/// \brief Augments the (char *) shader source code (as a string_view) with
+/// (optional) way to map the line in error with an identifier.
+///
+/// The identifier might be a fixed string, or a mapping constructed by preprocessing includes.
 struct ShaderSourceView
 {
     /*implicit*/ ShaderSourceView(const std::string & aSource, std::string aIdentifier = "") :
@@ -115,10 +30,7 @@ struct ShaderSourceView
         mIdentifier{std::move(aIdentifier)}
     {}
 
-    /*implicit*/ ShaderSourceView(const ShaderSource & aShaderSource) :
-        mSource{aShaderSource.mSource},
-        mMap{&aShaderSource.getSourceMap()}
-    {}
+    /*implicit*/ ShaderSourceView(const ShaderSource & aShaderSource);
 
     operator std::string_view () const
     { return mSource; }
